Move the query tests to new class.
