+++ test/edu/ucsb/nceas/metacat/dataone/MNodeQueryTest.java
@@ -1,0 +1,203 @@
+/**
+ *  '$RCSfile$'
+ *  Copyright: 2010 Regents of the University of California and the
+ *              National Center for Ecological Analysis and Synthesis
+ *  Purpose: To test the Access Controls in metacat by JUnit
+ *
+ *   '$Author:$'
+ *     '$Date:$'
+ * '$Revision:$'
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ */
+
+package edu.ucsb.nceas.metacat.dataone;
+
+
+
+
+import edu.ucsb.nceas.metacat.IdentifierManager;
+import edu.ucsb.nceas.metacat.dataone.CNodeService;
+import edu.ucsb.nceas.metacat.dataone.MNodeService;
+import edu.ucsb.nceas.metacat.properties.PropertyService;
+import edu.ucsb.nceas.metacat.properties.SkinPropertyService;
+import edu.ucsb.nceas.metacat.service.ServiceService;
+import edu.ucsb.nceas.utilities.IOUtil;
+import gov.loc.repository.bagit.Bag;
+import gov.loc.repository.bagit.BagFactory;
+import gov.loc.repository.bagit.BagFile;
+import gov.loc.repository.bagit.Manifest;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.UnsupportedEncodingException;
+import java.math.BigInteger;
+import java.net.URL;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Vector;
+
+import junit.framework.Test;
+import junit.framework.TestSuite;
+
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
+import org.dataone.client.v2.formats.ObjectFormatCache;
+import org.dataone.exceptions.MarshallingException;
+import org.dataone.configuration.Settings;
+import org.dataone.ore.ResourceMapFactory;
+import org.dataone.service.util.Constants;
+import org.dataone.service.util.TypeMarshaller;
+import org.dataone.service.exceptions.IdentifierNotUnique;
+import org.dataone.service.exceptions.InsufficientResources;
+import org.dataone.service.exceptions.InvalidRequest;
+import org.dataone.service.exceptions.InvalidSystemMetadata;
+import org.dataone.service.exceptions.InvalidToken;
+import org.dataone.service.exceptions.NotAuthorized;
+import org.dataone.service.exceptions.NotFound;
+import org.dataone.service.exceptions.NotImplemented;
+import org.dataone.service.exceptions.ServiceFailure;
+import org.dataone.service.exceptions.SynchronizationFailed;
+import org.dataone.service.exceptions.UnsupportedType;
+import org.dataone.service.types.v1.AccessPolicy;
+import org.dataone.service.types.v1.AccessRule;
+import org.dataone.service.types.v1.Checksum;
+import org.dataone.service.types.v1.DescribeResponse;
+import org.dataone.service.types.v1.Event;
+import org.dataone.service.types.v1.ObjectFormatIdentifier;
+import org.dataone.service.types.v1.Identifier;
+import org.dataone.service.types.v2.Log;
+import org.dataone.service.types.v2.Node;
+import org.dataone.service.types.v2.OptionList;
+import org.dataone.service.types.v1.NodeReference;
+import org.dataone.service.types.v1.ObjectFormatIdentifier;
+import org.dataone.service.types.v1.ObjectList;
+import org.dataone.service.types.v1.Permission;
+import org.dataone.service.types.v1.Person;
+import org.dataone.service.types.v1.ReplicationPolicy;
+import org.dataone.service.types.v1.Session;
+import org.dataone.service.types.v1.Subject;
+import org.dataone.service.types.v1.SubjectInfo;
+import org.dataone.service.types.v1.util.ChecksumUtil;
+import org.dataone.service.types.v2.SystemMetadata;
+import org.dspace.foresite.ResourceMap;
+import org.junit.After;
+import org.junit.Before;
+
+/**
+ * A JUnit test to exercise the Metacat Member Node  query service implementation.
+ * @author cjones
+ *
+ */
+public class MNodeQueryTest extends D1NodeServiceTest {
+
+    private static String unmatchingEncodingFilePath = "test/incorrect-encoding-declaration.xml";
+  /**
+   * Set up the test fixtures
+   * 
+   * @throws Exception
+   */
+  @Before
+  public void setUp() throws Exception {
+    super.setUp();
+    // set up the configuration for d1client
+    Settings.getConfiguration().setProperty("D1Client.cnClassName", MockCNode.class.getName());
+  }
+
+  /**
+   * Remove the test fixtures
+   */
+  @After
+  public void tearDown() {
+  }
+  
+  /**
+   * Build the test suite
+   * @return
+   */
+  public static Test suite() {
+    
+    TestSuite suite = new TestSuite();
+    suite.addTest(new MNodeQueryTest("initialize"));
+    suite.addTest(new MNodeQueryTest("testQueryOfArchivedObjects"));
+    
+    return suite;
+    
+  }
+  
+  /**
+   * Constructor for the tests
+   * 
+   * @param name - the name of the test
+   */
+  public MNodeQueryTest(String name) {
+    super(name);
+    
+  }
+
+  /**
+   * Initial blank test
+   */
+  public void initialize() {
+    assertTrue(1 == 1);
+    
+  }
+
+    public void testQueryOfArchivedObjects() throws Exception {
+        Session session = getTestSession();
+        Identifier guid = new Identifier();
+        guid.setValue("testUpdate." + System.currentTimeMillis());
+        InputStream object = new ByteArrayInputStream("test".getBytes("UTF-8"));
+        SystemMetadata sysmeta = createSystemMetadata(guid, session.getSubject(), object);
+        Identifier pid = MNodeService.getInstance(request).create(session, guid, object, sysmeta);
+        Thread.sleep(30000);
+        String query = "q=id:"+guid.getValue();
+        InputStream stream = MNodeService.getInstance(request).query(session, "solr", query);
+        String resultStr = IOUtils.toString(stream, "UTF-8");
+        System.out.println("the guid is "+guid.getValue());
+        System.out.println("the string is +++++++++++++++++++++++++++++++++++\n"+resultStr);
+        assertTrue(resultStr.contains("<str name=\"id\">"+guid.getValue()+"</str>"));
+        assertTrue(resultStr.contains("<bool name=\"archived\">false</bool>"));
+        
+        MNodeService.getInstance(request).archive(session, guid);
+        SystemMetadata result = MNodeService.getInstance(request).getSystemMetadata(session, guid);
+        Thread.sleep(30000);
+        stream = MNodeService.getInstance(request).query(session, "solr", query);
+        resultStr = IOUtils.toString(stream, "UTF-8");
+        assertTrue(!resultStr.contains("<str name=\"id\">"+guid.getValue()+"</str>"));
+        assertTrue(!resultStr.contains("<bool name=\"archived\">false</bool>"));
+        
+        query = "q=id:"+guid.getValue()+"&archived=archived:true";
+        stream = MNodeService.getInstance(request).query(session, "solr", query);
+        resultStr = IOUtils.toString(stream, "UTF-8");
+        System.out.println("the guid is "+guid.getValue());
+        System.out.println("the string is +++++++++++++++++++++++++++++++++++\n"+resultStr);
+        assertTrue(resultStr.contains("<str name=\"id\">"+guid.getValue()+"</str>"));
+        assertTrue(resultStr.contains("<bool name=\"archived\">true</bool>"));
+        
+    }
+
+}
